added impl package to service and dao, moved there all implementations; made little fixes in README.md.
