+++ src/main/java/taxi/service/AuthenticationService.java
@@ -1,4 +1,4 @@
-package taxi.service.authentication;
+package taxi.service;
 
 import taxi.exception.AuthenticationException;
 import taxi.model.Driver;
