+++ src/main/java/taxi/service/impl/DriverServiceImpl.java
@@ -1,4 +1,4 @@
-package taxi.service;
+package taxi.service.impl;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,6 +6,7 @@ import taxi.dao.DriverDao;
 import taxi.lib.Inject;
 import taxi.lib.Service;
 import taxi.model.Driver;
+import taxi.service.DriverService;
 
 @Service
 public class DriverServiceImpl implements DriverService {
