+++ src/main/java/taxi/service/impl/ManufacturerServiceImpl.java
@@ -1,10 +1,11 @@
-package taxi.service;
+package taxi.service.impl;
 
 import java.util.List;
 import taxi.dao.ManufacturerDao;
 import taxi.lib.Inject;
 import taxi.lib.Service;
 import taxi.model.Manufacturer;
+import taxi.service.ManufacturerService;
 
 @Service
 public class ManufacturerServiceImpl implements ManufacturerService {
